menambahkan middleware dan mengedit tampilan menjadi dinamis

+++ routes/web.php
@@ -15,13 +15,13 @@ Route::post('/register', [AuthController::class, 'register']);
 
 Route::get('/dashboard', function () {
     return view('pages.dashboard');
-})->name('dashboard');
+})->middleware('role:Admin,User');
 
 //resident hak akses admin
-Route::get('/resident', [ResidentController::class, 'index']);
-Route::get('/resident/create', [ResidentController::class, 'create']);
-Route::get('/resident/{id}', [ResidentController::class, 'edit']);
-Route::post('/resident', [ResidentController::class, 'store']);
-Route::put('/resident/{id}', [ResidentController::class, 'update']);
-Route::delete('/resident/{id}', [ResidentController::class, 'destroy']);
+Route::get('/resident', [ResidentController::class, 'index'])->middleware('role:Admin');
+Route::get('/resident/create', [ResidentController::class, 'create'])->middleware('role:Admin');
+Route::get('/resident/{id}', [ResidentController::class, 'edit'])->middleware('role:Admin');
+Route::post('/resident', [ResidentController::class, 'store'])->middleware('role:Admin');
+Route::put('/resident/{id}', [ResidentController::class, 'update'])->middleware('role:Admin');
+Route::delete('/resident/{id}', [ResidentController::class, 'destroy'])->middleware('role:Admin');
